Moved interfaces to different target
