+++ src/engine/source/fs/include/fs/xz/fileDataProvider.hpp
@@ -15,7 +15,7 @@
 #include <fstream>
 #include <vector>
 
-#include "iDataProvider.hpp"
+#include <fs/iDataProvider.hpp>
 
 namespace fs::xz
 {
