+++ src/engine/source/fs/include/fs/xz/stringDataProvider.hpp
@@ -13,7 +13,7 @@
 
 #include <string>
 
-#include "iDataProvider.hpp"
+#include <fs/iDataProvider.hpp>
 
 namespace fs::xz
 {
