+++ src/engine/source/fs/include/fs/xz/vectorDataCollector.hpp
@@ -15,7 +15,7 @@
 #include <cstdint>
 #include <vector>
 
-#include "iDataCollector.hpp"
+#include <fs/iDataCollector.hpp>
 
 namespace fs::xz
 {
