+++ src/engine/source/fs/include/fs/xz/vectorDataProvider.hpp
@@ -15,7 +15,7 @@
 #include <cstdint>
 #include <vector>
 
-#include "iDataProvider.hpp"
+#include <fs/iDataProvider.hpp>
 
 namespace fs::xz
 {
